refactor: AnalysisResultDTO를 vo 디렉토리로 옮긴 후 이름 변경
